+++ pub/unrouter/lib/unrouter.dart
@@ -8,7 +8,4 @@ export 'src/core/route_data.dart';
 export 'src/core/route_definition.dart';
 export 'src/runtime/unrouter.dart';
 export 'src/runtime/runtime_state.dart'
-    show
-        UnrouterResolutionState,
-        UnrouterStateSnapshot,
-        UnrouterStateTimelineEntry;
+    show UnrouterResolutionState, UnrouterStateSnapshot;